refactor: use nested controller paths for cleaner routes
- ConsensusController: @Controller('tasks/consensus')
- EscalationsController: @Controller('tasks/escalations')
- TaskTemplatesController: @Controller('tasks/templates')
- TaskRoutingController: @Controller('tasks/routing') [new]

Task-specific routes (escalate, create-template, candidates, auto-assign)
remain in TasksController since they operate on task :id.

This follows NestJS convention and eliminates repeated prefixes in decorators.
