+++ apps/api/src/tasks/task-templates.controller.ts
@@ -1,24 +1,24 @@
-import { Body, Controller, Delete, Get, Param, Post, Query } from "@nestjs/common";
+import { Body, Controller, Delete, Get, Param, Post } from "@nestjs/common";
 
 import { CurrentAgent, type AuthenticatedAgent } from "../auth";
 
-import { TaskRoutingService } from "./task-routing.service";
 import { TaskTemplatesService, type CreateTemplateDto, type InstantiateTemplateDto } from "./task-templates.service";
 
-@Controller("tasks")
+/**
+ * Handles task template CRUD operations.
+ * Task-specific template routes (create-template from task) are in TasksController.
+ */
+@Controller("tasks/templates")
 export class TaskTemplatesController {
-  constructor(
-    private readonly templatesService: TaskTemplatesService,
-    private readonly routingService: TaskRoutingService,
-  ) {}
+  constructor(private readonly templatesService: TaskTemplatesService) {}
 
-  @Get("templates")
+  @Get()
   async getTemplates(@CurrentAgent() agent: AuthenticatedAgent) {
     const templates = await this.templatesService.getTemplates(agent.orgId);
     return { data: templates };
   }
 
-  @Post("templates")
+  @Post()
   async createTemplate(
     @CurrentAgent() agent: AuthenticatedAgent,
     @Body() dto: CreateTemplateDto,
@@ -31,7 +31,7 @@ export class TaskTemplatesController {
     return { data: template };
   }
 
-  @Get("templates/:templateId")
+  @Get(":templateId")
   async getTemplate(
     @CurrentAgent() agent: AuthenticatedAgent,
     @Param("templateId") templateId: string,
@@ -40,7 +40,7 @@ export class TaskTemplatesController {
     return { data: template };
   }
 
-  @Delete("templates/:templateId")
+  @Delete(":templateId")
   async deleteTemplate(
     @CurrentAgent() agent: AuthenticatedAgent,
     @Param("templateId") templateId: string,
@@ -49,7 +49,7 @@ export class TaskTemplatesController {
     return { message: "Template deleted" };
   }
 
-  @Post("templates/instantiate")
+  @Post("instantiate")
   async instantiateTemplate(
     @CurrentAgent() agent: AuthenticatedAgent,
     @Body() dto: InstantiateTemplateDto,
@@ -60,62 +60,5 @@ export class TaskTemplatesController {
       dto,
     );
     return { data: tasks };
-  }
-
-  @Post(":id/create-template")
-  async createTemplateFromTask(
-    @CurrentAgent() agent: AuthenticatedAgent,
-    @Param("id") id: string,
-    @Body() body: { name: string },
-  ) {
-    const template = await this.templatesService.createFromTask(
-      agent.orgId,
-      agent.id,
-      id,
-      body.name,
-    );
-    return { data: template };
-  }
-
-  @Get(":id/candidates")
-  async findCandidates(
-    @CurrentAgent() agent: AuthenticatedAgent,
-    @Param("id") id: string,
-    @Query("minCoverage") minCoverage?: string,
-    @Query("maxResults") maxResults?: string,
-  ) {
-    const result = await this.routingService.findCandidates(agent.orgId, id, {
-      minCoverage: minCoverage ? parseInt(minCoverage, 10) : undefined,
-      maxResults: maxResults ? parseInt(maxResults, 10) : undefined,
-    });
-    return { data: result };
-  }
-
-  @Post(":id/auto-assign")
-  async autoAssign(
-    @CurrentAgent() agent: AuthenticatedAgent,
-    @Param("id") id: string,
-    @Body() body: { minCoverage?: number; excludeAgentIds?: string[] },
-  ) {
-    const result = await this.routingService.autoAssign(agent.orgId, agent.id, id, {
-      minCoverage: body.minCoverage,
-      excludeAgentIds: body.excludeAgentIds,
-    });
-    return { data: result };
-  }
-
-  @Get("routing/suggest")
-  async suggestAgents(
-    @CurrentAgent() agent: AuthenticatedAgent,
-    @Query("capabilities") capabilities: string,
-    @Query("limit") limit?: string,
-  ) {
-    const capList = capabilities.split(",").map((c) => c.trim()).filter(Boolean);
-    const suggestions = await this.routingService.suggestAgents(
-      agent.orgId,
-      capList,
-      limit ? parseInt(limit, 10) : 5,
-    );
-    return { data: suggestions };
   }
 }
